+++ files/gitlab-cookbooks/crond/attributes/default.rb
@@ -1,0 +1,3 @@
+default['crond']['enable'] = false
+default['crond']['log_directory'] = '/var/log/gitlab/crond'
+default['crond']['cron_d'] = '/var/opt/gitlab/crond'