+++ files/gitlab-cookbooks/letsencrypt/recipes/enable.rb
@@ -31,9 +31,27 @@ end
 
 include_recipe "letsencrypt::#{node['letsencrypt']['authorization_method']}_authorization"
 
+if node['letsencrypt']['auto_renew']
+  include_recipe "crond::enable"
+
+  # We seed with the sha1 of the hostname, so we'll default with the same minute
+  # until external_url changes
+  chosen_minute = Random.new(Digest::SHA1.hexdigest(site.host).hex).rand(60)
+
+  crond_job 'letsencrypt-renew' do
+    user "root"
+    hour node['letsencrypt']['auto_renew_hour']
+    minute node['letsencrypt']['auto_renew_minute'] || chosen_minute
+    day_of_month node['letsencrypt']['auto_renew_day_of_month']
+    command "/opt/gitlab/bin/gitlab-ctl renew-le-certs"
+  end
+end
+
 ruby_block 'display_le_message' do
   block do
-    LoggingHelper.warning("Let's Encrypt integration does not setup any automatic renewal. Please see https://docs.gitlab.com/omnibus/settings/ssl.html#lets-encrypt-integration for more information")
+    unless node['letsencrypt']['auto_renew']
+      LoggingHelper.warning("Let's Encrypt integration does not setup any automatic renewal. Please see https://docs.gitlab.com/omnibus/settings/ssl.html#lets-encrypt-integration for more information")
+    end
   end
   action :nothing
 end