+++ spec/fixtures/cookbooks/test_crond_job/recipes/delete.rb
@@ -1,0 +1,3 @@
+crond_job 'delete' do
+  action :delete
+end