+++ spec/fixtures/cookbooks/test_crond_job/recipes/minimal.rb
@@ -1,0 +1,4 @@
+crond_job 'minimal' do
+  user "rspec"
+  command "echo 'Hello world'"
+end