+++ CHANGELOG.md
@@ -31,6 +31,9 @@ omnibus-gitlab repository.
 - Create shared/{tmp,cache} directories when manage-storage-directories is enabled
 - Render Gitaly logging_ruby_sentry_dsn config option
 - Fix bug in error reporting of `gitlab-ctl renew-le-certs` #3389
+- Adds go-crond to the package #3251
+- Auto-renew LetsEncrypt with go-crond #3251
+
 
 10.6.4